# backend/diary/openapi.yaml
openapi: 3.1.0
info:
  title: Dopa optional photo diary (local development)
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8787/v1
security:
  - session: []
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
...
  schemas:
...
    Error:
      type: object
      properties:
        code: { type: string }
      required: [code]
      additionalProperties: false
  responses:
    error:
      description: Stable error code, no payloads or exception details
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /sessions:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [offsetMinutes, consentVersion]
              properties:
                offsetMinutes: { type: integer, minimum: -720, maximum: 840 }
                consentVersion: { const: photo-transfer-v1 }
      responses:
        '201':
          description: Anonymous token; store securely, never log
          content:
            application/json:
              schema:
                type: object
                properties: { token: { type: string, pattern: '^[a-f0-9]{64}$' } }
        default: { $ref: '#/components/responses/error' }
  /session:
    delete:
      responses:
        '200': { description: Session and all temporary images deleted }
        default: { $ref: '#/components/responses/error' }
...